fix(build): cast gate-crossing size to size_t for the 32-bit Windows build
register_gate_crossing's size parameter is size_t (32-bit on the Windows
x86 host build). run_simulated_code_size is a plain uint64_t variable
(not a compile-time constant MSVC can prove fits in 32 bits, unlike the
old inline constexpr-subtraction expression it replaced), so MSVC's /WX
promoted the resulting C4244 narrowing-conversion warning to a hard error,
breaking the Windows x86 build in 5e54f810.

Gate-crossing spans are always a few hundred bytes at most, so the
truncation is never actually reachable - just needs an explicit cast to
satisfy the warning. Verified locally (macOS build, which doesn't hit this
32-bit-size_t warning): native test-sample.exe and FEX wow64-test-sample.exe
both still exit 0.

+++ src/windows-emulator/module/module_manager.cpp
@@ -921,7 +921,7 @@ namespace sogen
                     run_simulated_code_size = dispatch_start - *scanned;
                 }
 
-                emu_ptr->register_gate_crossing(run_simulated_code_address, run_simulated_code_size,
+                emu_ptr->register_gate_crossing(run_simulated_code_address, static_cast<size_t>(run_simulated_code_size),
                                                 x86_64_emulator::gate_crossing_kind::wow64_run_simulated_code);
 
                 // NtContinue never returns through the normal syscall path - the real kernel restores
